refactor: extract BaseCacheContext ABC shared by GPU and CPU cache contexts
GPUCacheContext and CpuCacheContext were duck-typing through a bunch of
shared properties and methods. Extract them into a BaseCacheContext ABC
with class-level annotations so both type-checkers and humans can see
the contract clearly.
