+++ lmcache/v1/platform/base_cache_context.py
@@ -1,0 +1,305 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Abstract base class for platform cache contexts.
+
+Defines the common interface shared by :class:`GPUCacheContext` and
+:class:`CpuCacheContext`.  Concrete subclasses provide
+device-specific implementations of stream / buffer / copy primitives
+while the base class owns layout-agnostic helpers (shape calculation,
+status reporting, block-ID staging).
+"""
+
+# Future
+from __future__ import annotations
+
+# Standard
+from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING, Any
+import array
+
+# Third Party
+import torch
+
+# First Party
+from lmcache.v1.gpu_connector.utils import (
+    get_attention_backend,
+    get_concrete_engine_kv_shape_from_shape_desc,
+    get_engine_kv_shape_description,
+    is_mla,
+)
+from lmcache.v1.kv_layer_groups import KVLayerGroupsManager
+
+if TYPE_CHECKING:
+    # First Party
+    import lmcache.c_ops as lmc_ops
+
+
+class BaseCacheContext(ABC):
+    """Abstract base for GPU and CPU cache contexts.
+
+    Subclasses must set the following instance attributes in
+    :meth:`__init__` and the class-level annotations below keep
+    static analysis happy without polluting the ABC with a
+    concrete :meth:`__init__`.
+    """
+
+    # -- Set by subclass __init__ (annotations silence IDE warnings) --
+
+    _engine_kv_format: Any
+    device_: torch.device
+    kv_caches_: list[torch.Tensor]
+    num_layers_: int
+    num_blocks_: int
+    is_mla_: bool
+    kv_layer_groups_manager_: KVLayerGroupsManager
+    block_ids_buffer_: torch.Tensor
+    lmcache_tokens_per_chunk: int
+
+    # ------------------------------------------------------------------
+    # Abstract -- subclasses MUST implement
+    # ------------------------------------------------------------------
+
+    @property
+    @abstractmethod
+    def dtype(self) -> torch.dtype:
+        """Returns the dtype of the KV cache tensors."""
+        ...
+
+    @property
+    @abstractmethod
+    def stream(self) -> Any:
+        """Returns the device-specific stream for async operations."""
+        ...
+
+    @property
+    @abstractmethod
+    def cupy_stream(self) -> Any:
+        """Returns the cupy ExternalStream wrapping *stream*."""
+        ...
+
+    @property
+    @abstractmethod
+    def max_batch_size(self) -> int:
+        """Returns the maximum number of concurrent batches."""
+        ...
+
+    @abstractmethod
+    def close(self) -> None:
+        """Release device-specific resources (GDS staging buffers, etc.)."""
+        ...
+
+    @abstractmethod
+    def get_kernel_group_kv_pointers(self, kernel_group_idx: int) -> torch.Tensor:
+        """Returns the KV-cache pointer tensor for *kernel_group_idx*."""
+        ...
+
+    @abstractmethod
+    def get_temp_kernel_group_buffer(
+        self, batch_idx: int, kernel_group_idx: int
+    ) -> torch.Tensor:
+        """Returns a typed temp-buffer view for a (batch, kernel-group)
+        pair."""
+        ...
+
+    @abstractmethod
+    def get_temp_object_group_buffer(
+        self, batch_idx: int, object_group_idx: int
+    ) -> torch.Tensor:
+        """Returns a flat uint8 temp-buffer view for a (batch, object-group)
+        pair."""
+        ...
+
+    @abstractmethod
+    def get_kernel_group_shape_dtype(
+        self,
+        num_tokens: int,
+        kernel_group_idx: int,
+    ) -> tuple[torch.Size, torch.dtype]:
+        """Returns ``(shape, dtype)`` for *kernel_group_idx*."""
+        ...
+
+    @abstractmethod
+    def cache_size_per_token(self) -> int:
+        """Returns cache size per logical token in bytes (all groups)."""
+        ...
+
+    # ------------------------------------------------------------------
+    # Concrete -- shared implementations
+    # ------------------------------------------------------------------
+
+    @property
+    def engine_kv_format_(self) -> Any:
+        """Returns the EngineKVFormat enum value."""
+        return self._engine_kv_format
+
+    @property
+    def device(self) -> torch.device:
+        """Returns the device where KV-cache tensors live."""
+        return self.device_
+
+    @property
+    def kv_tensors(self) -> list[torch.Tensor]:
+        """Returns the list of per-layer KV cache tensors."""
+        return self.kv_caches_
+
+    @property
+    def num_layers(self) -> int:
+        """Returns the number of layers in the model."""
+        return self.num_layers_
+
+    @property
+    def num_blocks(self) -> int:
+        """Returns the number of blocks in the KV cache."""
+        return self.num_blocks_
+
+    @property
+    def is_mla(self) -> bool:
+        """Returns whether the model uses MLA."""
+        return self.is_mla_
+
+    @property
+    def hidden_dim_sizes(self) -> list[int]:
+        """Returns hidden dimension sizes per KV layer group."""
+        return [
+            group.hidden_dim_size
+            for group in self.kv_layer_groups_manager_.kv_layer_groups
+        ]
+
+    @property
+    def kv_layer_groups_manager(self) -> KVLayerGroupsManager:
+        """Returns the KV layer groups manager."""
+        return self.kv_layer_groups_manager_
+
+    def calculate_num_blocks(self, num_tokens: int, kernel_group_idx: int) -> int:
+        """Calculate the number of blocks for *num_tokens* in a kernel
+        group."""
+        return self.kv_layer_groups_manager_.calculate_num_blocks(
+            kernel_group_idx, num_tokens
+        )
+
+    def get_shape_desc(self, group_idx: int) -> "lmc_ops.PageBufferShapeDesc":
+        """Returns the PageBufferShapeDesc for *group_idx*."""
+        return self.kv_layer_groups_manager_.get_shape_desc(group_idx)
+
+    def get_slots_per_chunk_in_sw(self, kernel_group_idx: int) -> int:
+        """Returns the number of slots per lmcache chunk for D/H
+        transfer."""
+        return self.kv_layer_groups_manager_.get_slots_per_chunk_in_sw(kernel_group_idx)
+
+    def get_kv_buffer_shape(
+        self, logical_num_tokens: int, group_idx: int = 0
+    ) -> torch.Size:
+        """Returns the KV buffer shape for *logical_num_tokens*."""
+        group = self.kv_layer_groups_manager_.kv_layer_groups[group_idx]
+        compress_ratio = group.tokens_per_block // group.slots_per_block
+        if logical_num_tokens % compress_ratio != 0:
+            raise ValueError(
+                "logical_num_tokens (%d) is not a multiple of "
+                "compress_ratio (%d) for group %d"
+                % (logical_num_tokens, compress_ratio, group_idx)
+            )
+        num_slots = logical_num_tokens // compress_ratio
+        sd = group.shape_desc
+        return torch.Size(
+            (sd.kv_size, group.num_layers, num_slots, group.hidden_dim_size)
+        )
+
+    def copy_view_block_ids_to_gpu(
+        self, block_ids_per_group: list[list[int]]
+    ) -> list[torch.Tensor]:
+        """Copy per-group block IDs into the shared staging buffer.
+
+        Returns one non-overlapping view per LMCache group.
+        """
+        offsets = [0]
+        flat: array.array = array.array("l")
+        for view_block_ids in block_ids_per_group:
+            flat.extend(view_block_ids)
+            offsets.append(len(flat))
+
+        total = offsets[-1]
+        if total > self.block_ids_buffer_.shape[0]:
+            raise ValueError(
+                "block ID total %d exceeds the pre-allocated buffer "
+                "size %d" % (total, self.block_ids_buffer_.shape[0])
+            )
+        if total:
+            cpu_tensor = torch.frombuffer(flat, dtype=torch.long)
+            self.block_ids_buffer_[:total].copy_(cpu_tensor, non_blocking=True)
+
+        return [
+            self.block_ids_buffer_[offsets[i] : offsets[i + 1]]
+            for i in range(len(block_ids_per_group))
+        ]
+
+    # ------------------------------------------------------------------
+    # Derived properties (pure helpers)
+    # ------------------------------------------------------------------
+
+    @property
+    def engine_kv_shape(self) -> str:
+        """Returns the symbolic KV cache layout description."""
+        return get_engine_kv_shape_description(self._engine_kv_format)
+
+    @property
+    def attention_backend(self) -> str:
+        """Returns the attention backend name."""
+        return get_attention_backend(self._engine_kv_format)
+
+    @property
+    def concrete_engine_kv_shape(self) -> str:
+        """Returns the engine KV shape with actual numeric values."""
+        group = self.kv_layer_groups_manager_.kv_layer_groups[0]
+        return get_concrete_engine_kv_shape_from_shape_desc(
+            group.shape_desc, self._engine_kv_format
+        )
+
+    # ------------------------------------------------------------------
+    # Shared report_status
+    # ------------------------------------------------------------------
+
+    def report_status(self) -> dict:
+        """Return this context's KV cache layout metadata."""
+        manager = self.kv_layer_groups_manager_
+        kernel_groups = manager.kernel_groups
+
+        kernel_group_to_object_group: dict[int, int] = {
+            kg_idx: og_idx
+            for og_idx, og in enumerate(manager.object_groups)
+            for kg_idx in og.kernel_group_indices
+        }
+
+        engine_kv_format = self._engine_kv_format
+        group_reports: list[dict] = []
+        for kernel_group_idx, group in enumerate(kernel_groups):
+            group_reports.append(
+                {
+                    "kernel_group_idx": kernel_group_idx,
+                    "engine_group_idx": group.engine_group_idx,
+                    "object_group_idx": kernel_group_to_object_group.get(
+                        kernel_group_idx, 0
+                    ),
+                    "num_layers": group.num_layers,
+                    "layer_indices": list(group.layer_indices),
+                    "tokens_per_block": group.tokens_per_block,
+                    "slots_per_block": group.slots_per_block,
+                    "dtype": str(group.dtype),
+                    "engine_kv_concrete_shape": (
+                        get_concrete_engine_kv_shape_from_shape_desc(
+                            group.shape_desc, engine_kv_format
+                        )
+                    ),
+                    "is_mla": is_mla(engine_kv_format),
+                    "engine_kv_format": engine_kv_format.name,
+                    "engine_kv_shape": get_engine_kv_shape_description(
+                        engine_kv_format
+                    ),
+                    "attention_backend": get_attention_backend(engine_kv_format),
+                }
+            )
+
+        return {
+            "num_layers": self.num_layers_,
+            "num_blocks": self.num_blocks_,
+            "cache_size_per_token": self.cache_size_per_token(),
+            "kernel_groups": group_reports,
+        }
